dashboard/dashboard_requirements.txt
streamlit
psycopg2-binary
pandas
sqlalchemy
python-dotenv
plotly